organize src folder
